+++ lib/src/theme/themes/g10/components/g10_content_switcher_theme_data.dart
@@ -6,8 +6,8 @@ class G10ContentSwitcherThemeData extends CarbonContentSwitcherThemeData {
   const G10ContentSwitcherThemeData()
       : super(
           contentSwitcherSelected: CarbonPalette.white,
-          contentSwitcherDivider: CarbonPalette.white,
-          contentSwitcherTextOnColor: CarbonPalette.white,
+          contentSwitcherDivider: CarbonPalette.gray30,
+          contentSwitcherTextOnColor: CarbonPalette.gray100,
           contentSwitcherBackground: CarbonPalette.gray20,
           contentSwitcherBackgroundHover: CarbonPalette.gray20Hover,
         );